CapitalResult
